morphology: remove --with-lib-path
This is needed for it to look in /usr/lib. This is needed
for gcc later, which needs to link against the system zlib, which
is installed in /usr/lib. Without it, it will not find it because
it only looks in *lib64 and /tools.

It used to coincidentally work before, since the stage 2 artifacts were
similar enough. This similarity broke building on ARM after shared
objects for zlib were introduced, since the stage 2 shared objects
were warped, and incompatible with correctly built shared objects.