morph: Do not attempt to fix specs
The script called stage2-eglibc-fix-specs would modify gcc's specs file
so that it would build things differently.

This was required, since it is difficult to alter the gcc command lines
to fix this, as built systems don't always make this easy.

This is no longer possible, since builds are now only allowed to write
to $chunk.build, $chunk.inst and /tmp.

However it is no longer necessary to do the fix here, since eglibc now
installs fixed specs.