Import binutils 2.25 tarball
The import was done by:
1) removing all tracked files
2) removing all .gitignore files
3) extracting the 2.25 tarball into the repo
4) reassembling patch e7277
5) creating a single commit with the result of the above operations

Removing all .gitignore files is crucial, otherwise the generated C
files by bison and flex will not be committed, causing binutils to not
compile on stage 3 (and a lot of pain tracking down why it doesn't, true
story).

Note that the morphologies files were not added back, as they are placed
in the definitions repo as well.